♻ refactor Factored out remaining Rx Tab components
- RxHistory
- RxOtc
- RxPillbox
- RxPrint
